Scaffold CLI structure for codex-tasks
