+++ src/worker/event_processor.rs
